CRUD BY ALPEN IS READY

+++ src/main/java/com/uas/pbo/repository/TransactionDetailRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uas.pbo.model.TransactionDetail;
 
 @Repository
-public interface TransactionDetailRepository extends JpaRepository<TransactionDetail, Long> {
+public interface TransactionDetailRepository extends JpaRepository<TransactionDetail, Integer> {
 
 }
